adds file to store modules which hold methods unique to specific repositories
