+++ lib/invoice_item_repository.rb
@@ -11,21 +11,8 @@ class InvoiceItemRepository
     create_repository(invoice_items, InvoiceItem)
   end
 
-  def find_all_by_item_id(ii)
-    @repository.values.find_all do |invoice_item|
-      invoice_item.item_id == ii
-    end
-  end
-
-  def find_all_by_invoice_id(ii)
-    @repository.values.find_all do |invoice_item|
-      invoice_item.invoice_id == ii
-    end
-  end
-
   def create(attributes)
     general_create(attributes, InvoiceItem)
   end
-
 
 end
