+++ lib/repository_helper_modules/helper.rb
@@ -1,0 +1,1 @@
+require_relative './invoice_item_helper'
